requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
pandas>=1.5.0
openpyxl>=3.1.0
xlwings>=0.30.0
scikit-learn>=1.2.0
pyyaml>=6.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
joblib>=1.2.0
python-dotenv>=1.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
